# docs/swagger.yaml
basePath: /api
definitions:
...
  dto.APIResponse:
    properties:
      data: {}
      message:
        type: string
      success:
        type: boolean
    type: object
...
info:
  contact: {}
  description: new-api AI 网关 —— 用户与账号相关接口(对外第三方)
  title: new-api 用户 API
  version: "1.0"
paths:
...
  /user/passkey:
    delete:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.APIResponse'
      security:
      - ApiKeyAuth: []
      summary: 解绑当前用户 Passkey
      tags:
      - 用户-passkey
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.APIResponse'
      security:
      - ApiKeyAuth: []
      summary: 查询当前用户 Passkey 绑定状态
      tags:
      - 用户-passkey
...
  /user/passkey/register/begin:
    post:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.APIResponse'
      security:
      - ApiKeyAuth: []
      summary: 启动 Passkey 注册
      tags:
      - 用户-passkey
  /user/passkey/register/finish:
    post:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.APIResponse'
      security:
      - ApiKeyAuth: []
      summary: 完成 Passkey 注册
      tags:
      - 用户-passkey
  /user/passkey/verify/begin:
    post:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.APIResponse'
      security:
      - ApiKeyAuth: []
      summary: 启动 Passkey 验证
      tags:
      - 用户-passkey
  /user/passkey/verify/finish:
    post:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.APIResponse'
      security:
      - ApiKeyAuth: []
      summary: 完成 Passkey 验证
      tags:
      - 用户-passkey
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"